Revert "stale asset materialization for schedules and sensors (#11516)"
This reverts commit 8848f6ebe8460874f353442712f5a22d718ad5ae while we sort out a fix for user code running on 1.0.5 and 1.0.6.

Test Plan: BK
