+++ python_modules/dagster/dagster/_daemon/sensor.py
@@ -34,7 +34,6 @@ from dagster._core.storage.pipeline_run import DagsterRun, DagsterRunStatus, Run
 from dagster._core.storage.tags import RUN_KEY_TAG, SENSOR_NAME_TAG
 from dagster._core.telemetry import SENSOR_RUN_CREATED, hash_name, log_action
 from dagster._core.workspace.context import IWorkspaceProcessContext
-from dagster._scheduler.stale import resolve_asset_selection
 from dagster._utils.error import SerializableErrorInfo, serializable_error_info_from_exc_info
 from dagster._utils.merger import merge_dicts
 
@@ -631,16 +630,6 @@ def _evaluate_sensor(
     )
 
     for run_request in sensor_runtime_data.run_requests:
-        asset_selection = resolve_asset_selection(workspace_process_context, run_request, external_sensor)  # type: ignore
-        if (
-            asset_selection is not None and len(asset_selection) == 0
-        ):  # asset selection is empty set after filtering for stale
-            continue
-        elif asset_selection is not None:
-            run_request = run_request.with_replaced_attrs(
-                asset_selection=asset_selection, stale_assets_only=False
-            )
-
         target_data: ExternalTargetData = check.not_none(
             external_sensor.get_target_data(run_request.job_name)
         )
