+++ python_modules/dagster/dagster_tests/scheduler_tests/test_scheduler_run.py
@@ -15,7 +15,6 @@ from dagster import (
     asset,
     define_asset_job,
     job,
-    materialize,
     op,
     repository,
     schedule,
@@ -450,25 +449,17 @@ default_config_schedule = ScheduleDefinition(
 
 @asset
 def asset1():
-    return "asset1"
+    ...
 
 
 @asset
-def asset2(asset1):
-    return asset1 + "asset2"
+def asset2():
+    ...
 
 
-asset_job = define_asset_job("asset_job")
-
-
-@schedule(job=asset_job, cron_schedule="@daily")
+@schedule(job=define_asset_job("asset_job"), cron_schedule="@daily")
 def asset_selection_schedule():
     return RunRequest(asset_selection=[asset1.key])
-
-
-@schedule(job=asset_job, cron_schedule="@daily")
-def stale_asset_selection_schedule():
-    return RunRequest(stale_assets_only=True)
 
 
 @repository
@@ -504,7 +495,6 @@ def the_repo():
         empty_schedule,
         [asset1, asset2],
         asset_selection_schedule,
-        stale_asset_selection_schedule,
     ]
 
 
@@ -2471,84 +2461,6 @@ def test_asset_selection(instance, workspace_context, external_repo, executor):
         assert run.asset_selection == {AssetKey("asset1")}
 
         validate_run_started(instance, run, execution_time=create_pendulum_time(2019, 2, 28))
-
-
-@pytest.mark.parametrize("executor", get_schedule_executors())
-def test_stale_asset_selection_never_materialized(
-    instance, workspace_context, external_repo, executor
-):
-    freeze_datetime = _get_stale_asset_selection_schedule_start()
-    external_schedule = external_repo.get_external_schedule("stale_asset_selection_schedule")
-
-    with pendulum.test(freeze_datetime):
-        instance.start_schedule(external_schedule)
-
-    # never materialized so all assets stale
-    freeze_datetime = freeze_datetime.add(seconds=2)
-    with pendulum.test(freeze_datetime):
-        evaluate_schedules(workspace_context, executor, pendulum.now("UTC"))
-        wait_for_all_runs_to_start(instance)
-        schedule_run = next(
-            (r for r in instance.get_runs() if r.pipeline_name == "asset_job"), None
-        )
-        assert schedule_run is not None
-        assert schedule_run.asset_selection == {AssetKey("asset1"), AssetKey("asset2")}
-        validate_run_started(
-            instance, schedule_run, execution_time=create_pendulum_time(2019, 2, 28)
-        )
-
-
-@pytest.mark.parametrize("executor", get_schedule_executors())
-def test_stale_asset_selection_empty(instance, workspace_context, external_repo, executor):
-    freeze_datetime = _get_stale_asset_selection_schedule_start()
-    external_schedule = external_repo.get_external_schedule("stale_asset_selection_schedule")
-
-    with pendulum.test(freeze_datetime):
-        instance.start_schedule(external_schedule)
-
-    materialize([asset1, asset2], instance=instance)
-
-    # assets previously materialized so we expect empy set
-    freeze_datetime = freeze_datetime.add(seconds=2)
-    with pendulum.test(freeze_datetime):
-        evaluate_schedules(workspace_context, executor, pendulum.now("UTC"))
-        wait_for_all_runs_to_start(instance)
-        schedule_run = next(
-            (r for r in instance.get_runs() if r.pipeline_name == "asset_job"), None
-        )
-        assert schedule_run is None
-
-
-@pytest.mark.parametrize("executor", get_schedule_executors())
-def test_stale_asset_selection_subset(instance, workspace_context, external_repo, executor):
-    freeze_datetime = _get_stale_asset_selection_schedule_start()
-    external_schedule = external_repo.get_external_schedule("stale_asset_selection_schedule")
-
-    with pendulum.test(freeze_datetime):
-        instance.start_schedule(external_schedule)
-
-    materialize([asset1], instance=instance)
-
-    # assets previously materialized so we expect empy set
-    freeze_datetime = freeze_datetime.add(seconds=2)
-    with pendulum.test(freeze_datetime):
-        evaluate_schedules(workspace_context, executor, pendulum.now("UTC"))
-        wait_for_all_runs_to_start(instance)
-        schedule_run = next(
-            (r for r in instance.get_runs() if r.pipeline_name == "asset_job"), None
-        )
-        assert schedule_run is not None
-        assert schedule_run.asset_selection == {AssetKey("asset2")}
-        validate_run_started(
-            instance, schedule_run, execution_time=create_pendulum_time(2019, 2, 28)
-        )
-
-
-def _get_stale_asset_selection_schedule_start():
-    return to_timezone(
-        create_pendulum_time(year=2019, month=2, day=27, hour=23, minute=59, second=59, tz="UTC"),
-        "US/Central",
-    )
 
 
 def test_settings():
